refacto(archi): move CV
