+++ src/CV/Domain/Project.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace App\Domain\CV;
+namespace App\CV\Domain;
 
-use App\Domain\CV\Repository\ProfessionalExperienceRepository;
+use App\CV\Domain\Repository\ProjectRepository;
 use App\Domain\Routing\ValueObject\RouteName;
 use App\Http\Cache\CacheableEntity;
 use App\Http\Cache\ValueObject\CacheItem;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
 
-#[ORM\Entity(repositoryClass: ProfessionalExperienceRepository::class)]
-#[ORM\Table(name: 'cv_professional_experience')]
+#[ORM\Entity(repositoryClass: ProjectRepository::class)]
+#[ORM\Table(name: 'cv_project')]
 #[ORM\HasLifecycleCallbacks]
-class ProfessionalExperience implements CacheableEntity
+class Project implements CacheableEntity
 {
     #[ORM\Id]
     #[ORM\GeneratedValue]
@@ -21,35 +20,16 @@ class ProfessionalExperience implements CacheableEntity
     private ?int $id = null;
 
     #[ORM\Column(length: 255)]
-    #[Assert\NotBlank()]
-    private ?string $company = null;
+    private ?string $name = null;
 
     #[ORM\Column(length: 255)]
-    #[Assert\NotBlank()]
-    #[Assert\Url()]
     private ?string $url = null;
 
-    #[ORM\Column(length: 255)]
-    #[Assert\NotBlank()]
-    private ?string $jobName = null;
-
     #[ORM\Column(type: Types::TEXT)]
-    #[Assert\NotBlank()]
     private ?string $description = null;
 
     #[ORM\Column(type: Types::DATE_IMMUTABLE)]
-    #[Assert\NotBlank()]
-    private ?\DateTimeImmutable $startDate = null;
-
-    #[ORM\Column(type: Types::DATE_IMMUTABLE, nullable: true)]
-    #[Assert\GreaterThan(propertyPath: 'startDate')]
-    private ?\DateTimeImmutable $endDate = null;
-
-    #[ORM\Column]
-    private ?\DateTimeImmutable $createdAt = null;
-
-    #[ORM\Column]
-    private ?\DateTimeImmutable $updatedAt = null;
+    private ?\DateTimeImmutable $date = null;
 
     /**
      * @var list<string>
@@ -57,21 +37,27 @@ class ProfessionalExperience implements CacheableEntity
     #[ORM\Column(type: Types::JSON, options: [
         'jsonb' => true,
     ])]
-    private array $badges = [];
+    private array $techStack = [];
+
+    #[ORM\Column]
+    private ?\DateTimeImmutable $createdAt = null;
+
+    #[ORM\Column]
+    private ?\DateTimeImmutable $updatedAt = null;
 
     public function getId(): ?int
     {
         return $this->id;
     }
 
-    public function getCompany(): ?string
+    public function getName(): ?string
     {
-        return $this->company;
+        return $this->name;
     }
 
-    public function setCompany(string $company): static
+    public function setName(string $name): static
     {
-        $this->company = $company;
+        $this->name = $name;
 
         return $this;
     }
@@ -88,18 +74,6 @@ class ProfessionalExperience implements CacheableEntity
         return $this;
     }
 
-    public function getJobName(): ?string
-    {
-        return $this->jobName;
-    }
-
-    public function setJobName(string $jobName): static
-    {
-        $this->jobName = $jobName;
-
-        return $this;
-    }
-
     public function getDescription(): ?string
     {
         return $this->description;
@@ -112,26 +86,14 @@ class ProfessionalExperience implements CacheableEntity
         return $this;
     }
 
-    public function getStartDate(): ?\DateTimeImmutable
+    public function getDate(): ?\DateTimeImmutable
     {
-        return $this->startDate;
+        return $this->date;
     }
 
-    public function setStartDate(\DateTimeImmutable $startDate): static
+    public function setDate(\DateTimeImmutable $date): static
     {
-        $this->startDate = $startDate;
-
-        return $this;
-    }
-
-    public function getEndDate(): ?\DateTimeImmutable
-    {
-        return $this->endDate;
-    }
-
-    public function setEndDate(?\DateTimeImmutable $endDate): static
-    {
-        $this->endDate = $endDate;
+        $this->date = $date;
 
         return $this;
     }
@@ -139,17 +101,17 @@ class ProfessionalExperience implements CacheableEntity
     /**
      * @return list<string>
      */
-    public function getBadges(): array
+    public function getTechStack(): array
     {
-        return $this->badges;
+        return $this->techStack;
     }
 
     /**
-     * @param array<string> $badges
+     * @param array<string> $techStack
      */
-    public function setBadges(array $badges): static
+    public function setTechStack(array $techStack): static
     {
-        $this->badges = array_values($badges);
+        $this->techStack = array_values($techStack);
 
         return $this;
     }
