+++ src/Http/Controller/Admin/CV/ProfessionalExperienceCrudController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controller\Admin\CV;
 
-use App\Domain\CV\ProfessionalExperience;
+use App\CV\Domain\ProfessionalExperience;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
 use EasyCorp\Bundle\EasyAdminBundle\Field\ArrayField;
